[py3k] object no longer has a default __lt__
In python 2:

Objects of different types except numbers are ordered by their type
names; objects of the same types that don?t support proper comparison
are ordered by their address.

In python 3:

Instances of a class cannot be ordered with respect to other instances
of the same class, or other types of object, unless the class defines
enough of the methods __lt__(), __le__(), __gt__(), and __ge__().

Related to #296242.