[fi] fix __eq__/__hash__ implementations
When redefining __eq__, we need to make sure __hash__ is consistent so
that two objects that compare equal will have the same hash.

Python 3 somewhat enforces that by defaulting __hash__ to None if __eq__
is defined.  We want these types to be hashable because we use them as
dictionary keys.

Related to #296242.